+++ src/gitlab/review.c
@@ -27,56 +27,23 @@
  * OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-#ifndef REVIEW_H
-#define REVIEW_H
+#include <ghcli/gitlab/review.h>
 
 #include <sn/sn.h>
 
-typedef struct ghcli_pr_review         ghcli_pr_review;
-typedef struct ghcli_pr_review_comment ghcli_pr_review_comment;
+size_t
+gitlab_review_get_reviews(
+    const char       *owner,
+    const char       *repo,
+    int               pr,
+    ghcli_pr_review **out)
+{
+    (void) owner;
+    (void) repo;
+    (void) pr;
 
-struct ghcli_pr_review_comment {
-    char *id;
-    char *author;
-    char *date;
-    char *diff;
-    char *path;
-    char *body;
-    int   original_position;
-};
+    warnx("Reviews are not yet supported on Gitlab");
+    *out = NULL;
 
-struct ghcli_pr_review {
-    char                    *id;
-    char                    *author;
-    char                    *date;
-    char                    *state;
-    char                    *body;
-    ghcli_pr_review_comment *comments;
-    size_t                   comments_size;
-};
-
-void ghcli_review_reviews_free(
-    ghcli_pr_review *it,
-    size_t size);
-
-void ghcli_review_comments_free(
-    ghcli_pr_review_comment *it,
-    size_t size);
-
-size_t ghcli_review_get_reviews(
-    const char *owner,
-    const char *repo,
-    int pr,
-    ghcli_pr_review **out);
-
-void ghcli_review_print_review_table(
-    FILE *,
-    ghcli_pr_review *,
-    size_t);
-
-void ghcli_review_print_comments(
-    FILE *out,
-    ghcli_pr_review_comment *comments,
-    size_t comments_size);
-
-#endif /* REVIEW_H */
+    return 0;
+}
